+++ src/features/exercise/ExerciseSlice.ts
@@ -1,10 +1,7 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit'
-
-import axios from 'axios'
-
+import {  createSlice, PayloadAction } from '@reduxjs/toolkit'
 import type { AppState } from '@app/store'
-import { ExerciseEntry, ExerciseTemplate, Program, Workout, WorkoutEntry } from '@prisma/client'
-import { LastWorkoutEntry } from '@server/getLastWorkoutOfType'
+import { ExerciseEntry, ExerciseTemplate, Workout, WorkoutEntry } from '@prisma/client'
+import {getExerciseTemplates, postExerciseEntries, getMorePreviousWorkouts, getWorkoutAsync} from '@features/exercise/thunks'
 
 //this holds the miscellaneous data about the workout that isn't tied to a specific exercise instead the entire workout in general
 export interface activeWorkoutInfo {
@@ -16,7 +13,7 @@ export interface activeWorkoutInfo {
 
 
 //combination of workout templates and workout entries, 
-interface WorkoutInfo extends Workout, Omit<WorkoutEntry, 'date|id'> {
+export interface WorkoutInfo extends Workout, Omit<WorkoutEntry, 'date|id'> {
 	prevDate: string
 	prevWorkoutId: number
 }
@@ -58,114 +55,9 @@ const initialState = {
 	workoutEntry: null as activeWorkoutInfo,
 }
 
-//get the list of workouts when we initialize the page , 
-//ie pull heavy, legs light etc and combine with the last workout of that type that was done
-export const getWorkoutAsync = createAsyncThunk(
-	'exercise/getWorkout',
-	async (_, { rejectWithValue }) => {
-		try {
 
-			const { id } = await axios.get('/api/program')
-				.then(d => d.data)
-				.catch(err => {
-					throw Error(err)
-				}) as Program
 
-			const workoutTemplates = await axios.get('/api/workout-template', { params: { programId: id } })
-				.then(r => r.data)
-				.catch(err => {
-					throw Error(err)
-				}) as Workout[]
 
-			const prevWorkouts = await Promise.all(workoutTemplates.map(workout => axios.get('/api/workout-entry', { params: { workoutType: workout.id, skip: 0 } })))
-				.then(list => list.map(item => item.data))
-				.catch(err => {
-					throw Error(err)
-				}) as WorkoutEntry[]
-
-			const data = workoutTemplates.map((item) => {
-				const prevWorkout = prevWorkouts.find(workout => workout.workoutTemplateId === item.id)
-				return {
-					...prevWorkout,
-					...item,
-					//label the workoutEntry id key as prevWorkoutId to avoid conflict with the workoutTemplate id key
-					prevWorkoutId: prevWorkout ? prevWorkout.id : null,
-				}
-			}) as WorkoutInfo[]
-
-			return data
-		}
-		catch (err) {
-			console.error(err)
-			return rejectWithValue('Unable to get workout async')
-		}
-	}
-)
-
-//given a workoutID get the template of exercises for that given workout
-export const getExerciseTemplates = createAsyncThunk(
-	'exercise/getExerciseTemplates',
-	async (id: number, { getState, rejectWithValue }) => {
-		try {
-			const { exercise: { workouts } } = getState() as AppState
-			const prevWorkoutID = workouts.find(workout => workout.id === id)?.prevWorkoutId
-
-			const [exercisesList, prevMeta, prevExercises] = await Promise.all([
-				axios.get('/api/exercise-templates', { params: { workoutId: id } }),
-				axios.get('/api/workout-entry', { params: { Id: prevWorkoutID } }),
-				axios.get('/api/exercise-entry', { params: { workoutId: prevWorkoutID } })
-			]).then(list => ([
-				list[0].data,
-				list[1].data as LastWorkoutEntry,
-				list[2].data as ExerciseEntry
-			]))
-
-			const previousWorkout: PreviousWorkout = prevWorkoutID && { ...prevMeta, date: prevMeta.date, exercises: prevExercises }
-
-			return {
-				exercises: exercisesList,
-				workoutId: id,
-				previousWorkout
-			}
-		} catch (err) {
-			console.error(err)
-			return rejectWithValue('Unable to get exercises')
-		}
-	}
-)
-
-/**
- * get more previous workouts of the same type to show older results
- */
-export const getMorePreviousWorkouts = createAsyncThunk(
-	'exercise/getMorePreviousWorkouts',
-	async (skipNum: number, { getState, rejectWithValue }) => {
-		try {
-			const { exercise: { activeWorkout } } = getState() as AppState
-			const workout = await axios.get('/api/workout-entry', { params: { workoutType: activeWorkout, skip: skipNum } })
-			if (!workout.data) return rejectWithValue('no readings found')
-			return workout.data
-		} catch (err) {
-			console.error('No Readings Found')
-			return rejectWithValue('No Readings Found')
-		}
-	}
-)
-
-export const postExerciseEntries = createAsyncThunk(
-	'exercise/postExerciseEntries',
-	async (_, { getState, rejectWithValue }) => {
-		const { exercise: { entries, workoutEntry } } = getState() as AppState
-		try { 
-			const response = await axios.post('/api/exercise-entry', { entries, workoutEntry })
-			return response.data
-		} catch (err) {
-			console.error(err)
-			return rejectWithValue('unable to post workouts , please try later')
-		}
-	}
-
-)
 
 export const exerciseSlice = createSlice({
 	name: 'exercise',
